+++ app/assets/scripts/minireel/app.js
@@ -358,10 +358,11 @@ function( angular , c6ui , c6log , c6State  , services          , tracker       
                 }
             };
 
-            function formatEmbed(template, data, processors) {
+            function formatEmbed(template, data, _processors) {
                 var repeatedMatcher = (/\|.+?\|/),
                     repeated = template.match(repeatedMatcher)[0].replace(/^.|.$/g, ''),
-                    bookends = template.split(repeatedMatcher);
+                    bookends = template.split(repeatedMatcher),
+                    processors = _processors || {};
 
                 function identity(arg) {
                     return arg;
@@ -411,30 +412,13 @@ function( angular , c6ui , c6log , c6State  , services          , tracker       
                 get: function() {
                     var minireel = $scope.minireel,
                         splash = minireel.data.splash,
-                        branding = minireel.data.branding,
                         isInline = MiniReelService.modeCategoryOf(minireel, categories)
                             .value === 'inline',
                         explicitDimensions = this.mode === 'custom';
 
-                    function shortcodeBase64(string) {
-                        var value = string.match(/"(.*?)"/)[1];
-
-                        return string.replace(value, 'data:text/plain;base64,' + btoa(value));
-                    }
-
-                    function scriptBase64(string) {
-                        var value = string.match(/"(.*?)"/)[1];
-
-                        return string
-                            .replace(/^data-/, 'data-:')
-                            .replace(value, btoa(value));
-                    }
-
                     var data = {
                         'exp': minireel.id,
-                        'title': minireel.data.title,
                         'splash': splash.theme + ':' + splash.ratio.replace('-', '/'),
-                        'branding': branding ? branding : false,
                         'width': explicitDimensions ? this.size.width : false,
                         'height': explicitDimensions ? this.size.height : false,
                         'preload': isInline
@@ -443,8 +427,6 @@ function( angular , c6ui , c6log , c6State  , services          , tracker       
                     switch (this.format) {
                     case 'shortcode':
                         return formatEmbed('[minireel version="1" |{attr}="{value}"|]', data, {
-                            title: shortcodeBase64,
-                            branding: shortcodeBase64,
                             preload: function(string) {
                                 return string + '="preload"';
                             }
@@ -452,11 +434,7 @@ function( angular , c6ui , c6log , c6State  , services          , tracker       
                     case 'script':
                         return formatEmbed(
                             '<script src="' + c6EmbedSrc + '" |data-{attr}="{value}"|></script>',
-                            data,
-                            {
-                                title: scriptBase64,
-                                branding: scriptBase64
-                            }
+                            data
                         );
                     }
                 }